Fixed VMR3PowerOn regression.
The problem was that the EMT outer loop wasn't checking for state 
changes in the init case (pUVM->pVM == NULL). So, if the VMR3PowerOn 
request was queued and executed by the initial VMR3ReqProcessU call
we wouldn't call EMR3ExecuteVM when the VM state changed to running.

svn:sync-xref-src-repo-rev: r27905