Main/DisplayImpl: Fix for access violation in case of multimonitor.
Display requests a resize of 4 framebuffers simultaneously. When one of framebuffers is resized, Display calls InvalidateAndUpdateEMT which is updating _all_ framebuffers but some of them are not yet finished with the resize. As a result, InvalidateAndUpdateEMT is assuming that all the framebuffers have right size, which is not true. Thus, crash while copying from source to destination recatangle .

svn:sync-xref-src-repo-rev: r83352