NAT: Make sure guest will get window updates by scheduling TF_DELACK on successful sowrite().
This code is a moral equivalent of calling tcp_output() for PRU_RCVD
in tcp_usrreq() of the real stack.  This is intended to be a
minimalistic fix to prevent guest from being stuck after we close
offered window.  Guest's persist timer eventually lets it continue,
but wastes a lot of wall time, making outbound transfers very slow.

svn:sync-xref-src-repo-rev: r94784