Main/DHCPServerImpl: for DHCP options keep not just the text representation, but also encoding of that text representation.
When XML settings are read, interpret old format, without explicit
"encoding" attribute, as legacy encoding where we are expected to know
the actual format of the option from the option code itself.  When
writing legacy options, write them in old format.

This is in preparation for using "de:ad:be:ef" hex-encoded option
values for non-standard options which format we can't know a priory.

svn:sync-xref-src-repo-rev: r98389