bugref:3809: Linux installer maintenance Try to fix dependencies on driver sources in Additions installer make file.
This change tries to duplicate the way dependencies on kernel module sources
are handled in the Linux host installer make file to ensure that the
Additions installer is always rebuilt when one of the source files changes.
There are lots of other clean-ups in the host make file which would be
worth copying, but I will put that off for whenever I need to change related
parts of the Additions make file.

svn:sync-xref-src-repo-rev: r117068