PS/2 mouse emulation: correct comment in test case and move test code to end.
bugref:8778: Double click not working with precision touchpad (user report)
Comments in the test case incorrectly stated that the interval between events
is 50ms.  In fact it is 10ms by default and adjustable by the guest operating
system.
Also move the test case code to the end of the file for added clarity.

svn:sync-xref-src-repo-rev: r117150