VBoxPci-linux.c: use kernel_write instead of vfs_write as of Linux 4.14. bugref:9057: VBoxPci: vfs_write() no longer exported by Linux kernel
As of Linux 4.14 vfs_write() is no longer exported for use by modules.  Other
users have been converted to use kernel_write(), which is a wrapper around
vfs_write() with some added boiler-plate calls to make it safe to use on
kernel addresses too.  It seems to me like we should probably already have
been using kernel_write(), but I prefer to leave that decision to someone who
knows the code, so this change switches over for 4.14 and later.

Fixes: ticketref:17267:comment:6

svn:sync-xref-src-repo-rev: r119144