slirp: bugref:9856 - in tcp_input() save optp and optlen in struct socket for the re-entry on connect dance.  Restore them on the second entry because the options are only parsed at that time.
I'm not sure why - this is code from the original slirp - but nothing
in the options parsing code needs getting the "remote address" as the
comment claims.  This bug is pretty corner case and has been
introduced in slirp changes to the BSD stack.  Opt for the
conservative fix.  The practical consequence is that we should now
respect the MSS that the guest advertises to us (ticketref:15256).

svn:sync-xref-src-repo-rev: r141292