Additions: Shared Clipboard: fix HTML content conversion, bugref:10160.
This commit applies a fix for HTML content handling for the case when
data is being copied from X11 host and pasted into Windows guest. We
now make sure that clipboard data is NULL-terminated string. If HTML
data is not wrapped into CF_HTML format, we now do so.

svn:sync-xref-src-repo-rev: r148773