IPRT/process-win: Fixed regression from r152441, leading to start Windows processes from services in session 0 instead of session >=1.
This was due to reworking the code to solely use the "W-variants" (Unicode) of the Windows Process32 APIs. rtProcWinFindTokenByProcess() now uses  PROCESSENTRY32W instead of PROCESSENTRY32, but it was forgotten to also change the dwSize member to sizeof(PROCESSENTRY32W). This in turn led to a debug assertion, showing a ERROR_BAD_LENGTH. The function itself, however, just returned false in release builds, which in turn messes up the process token handling.

svn:sync-xref-src-repo-rev: r161501