Devices/EFI/Firmware: Expose ACPI tables if available, bugref:10733
This requires changing the VBox descriptor slightly to accomodate for the fact that
the XSDP will be located at the end of the ACPI region but the start + size of the
whole region are required for setting up the identity mapped page tables and have this region
covered in the address space. Doesn't need bumping the version because older firmware never
exposed ACPI tables.


svn:sync-xref-src-repo-rev: r165165