bldprogs/VBoxEditElf: Experiment with creating stub linker libraries, bugref:10874
The tool parses a given shared library (.so), extracts the information relevant for linking
and writes the result into a new file which acts as a stub. This way .so files can get much smaller
and be distributed in an SDK style tools package without requiring a lot of disk space.


svn:sync-xref-src-repo-rev: r167897